requirements.txt
pandas
google-cloud-texttospeech
PyYAML
sortedcontainers
phonemizer
git+ssh://git@gitlab.cognitive-ml.fr:1022/htiteux/wuggy-ng.git#egg=wuggy_ng
git+ssh://git@github.com/bootphon/wordseg.git#egg=wordseg
git+ssh://git@github.com/cmusphinx/g2p-seq2seq.git#egg=g2p_seq2seq
# needed for habilis
tensorflow==1.8.0
tensorflow-gpu>=1.8.0